requirements.txt
# tensorflow==2.14.0
tensorflow-gpu==2.5.0
pandas 
gdown
dvc
mlflow==2.2.2
notebook
# numpy==1.22.4
# numpy
matplotlib
seaborn
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-PyYAML
scipy
Flask
Flask-Cors
# box
-e .